requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
convertdate==2.2.0
cryptography==3.3.1
decorator==4.4.2
dnspython==2.0.0
email-validator==1.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
holidays==0.10.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
korean-lunar-calendar==0.2.1
lxml==4.6.2
MarkupSafe==1.1.1
multitasking==0.0.9
numpy==1.19.4
pandas==1.1.5
pandas-datareader==0.9.0
pandas-market-calendars==1.6.1
psycopg2==2.8.6
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.20
PyMeeus==0.3.7
pymongo==3.11.2
python-dateutil==2.8.1
python-telegram-bot==13.1
pytz==2020.4
requests==2.25.0
rsa==4.6
simplejson==3.17.2
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.22
Telethon==1.18.2
toolz==0.11.1
tornado==6.1
trading-calendars==2.1.0
tzlocal==2.1
urllib3==1.26.2
Werkzeug==1.0.1
WTForms==2.3.3
yahoo-finance==1.4.0
yfinance==0.1.55
